only show modal button if GNMI_PROXY_URL is set don't show UNDERLAY vrf (since it's not part of network instances gnmi path) move component outside of interfaceconfig if we want to use it elsewhere
